Fix 0.1% intensity steps by passing Double and rounding correctly
setIntensity now takes Double instead of Int to preserve decimal
precision. Fixed floating point truncation in CCTProtocol by using
round() before Int conversion so 50.1% correctly maps to protocol
value 501 instead of potentially truncating to 500.

+++ FilmLightRemote/Sources/BLE/BLEManager.swift
@@ -304,15 +304,15 @@ class BLEManager: NSObject, ObservableObject {
         currentSaturation = Int(lightState.saturation)
     }
 
-    func setIntensity(_ percent: Int) {
-        currentIntensity = percent
+    func setIntensity(_ percent: Double) {
+        currentIntensity = Int(percent)
         currentMode = "cct"
         guard let peripheral = connectedPeripheral else { return }
 
         log("setIntensity(\(percent)%) target=0x\(String(format: "%04X", targetUnicastAddress))")
 
         // Sidus opcode 0x26 + CCTProtocol — controls intensity via CCT mode
-        let cctCmd = CCTProtocol(intensityPercent: Double(percent), cctKelvin: currentCCT)
+        let cctCmd = CCTProtocol(intensityPercent: percent, cctKelvin: currentCCT)
         let payload = cctCmd.getSendData()
         var accessMessage: [UInt8] = [0x26]
         accessMessage.append(contentsOf: payload)
